+++ scrapy/linkextractors/__init__.py
@@ -5,4 +5,5 @@ This package contains a collection of Link Extractors.
 
 For more info see docs/topics/link-extractors.rst
 """
+from .base import *
 from .lxmlhtml import LxmlLinkExtractor as LinkExtractor
