+++ scrapy/linkextractors/base.py
