+++ scrapy/linkextractors/lxmlhtml.py
@@ -11,7 +11,7 @@ from scrapy.selector import Selector
 from scrapy.link import Link
 from scrapy.utils.misc import arg_to_iter
 from scrapy.utils.python import unique as unique_list, str_to_unicode
-from scrapy.linkextractor import FilteringLinkExtractor
+from scrapy.linkextractors import FilteringLinkExtractor
 from scrapy.utils.response import get_base_url
